requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
lxml_html_clean>=0.1.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
instagrapi>=2.0.0
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.7.0
pyyaml>=6.0
aiohttp>=3.8.0
selenium>=4.15.0
webdriver-manager>=4.0.0
feedparser>=6.0.0
newspaper3k>=0.2.8
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0
fastapi>=0.100.0
uvicorn>=0.23.0
mcp>=1.0.0
